requirements.txt
# Orchestration & LLM
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langgraph>=0.0.10

# SQL Processing
sqlglot>=20.0.0

# Database Drivers & ORM
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0    # Target DB (PostgreSQL) Driver
cx_Oracle>=8.3.0          # Source DB (Oracle) Driver
# oracledb>=2.0.0         # (Option) cx_Oracle의 최신 대안, 필요시 교체 가능

# Utilities
pyyaml>=6.0
pydantic>=2.0.0
tqdm>=4.65.0              # CLI 진행률 표시용

pymssql>=2.2.0           # MS SQL Server (또는 pyodbc)
sqlalchemy-hana>=0.5.0   # SAP HANA
snowflake-sqlalchemy>=1.4.0 # Snowflake
pymongo>=4.3.0           # MongoDB (NoSQL)